+++ packages/enhanced/CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.14.4+1
+
+- Improve table support for wide columns (#1073)
+- BREAKING: remove `WidgetFactory.buildDivider` (#1075)
+- Make `align=center` work like `CENTER` tag (#1076)
+
 ## 0.14.3
 
 - Write migration docs from v0.10 to v0.14 (#1065)